Several Things
perp.js fixes a bug when the mysql server runs on windows-.
added a bit more logging information.
admin page is going to broadcast properly.
a fix for the app file from starter that will cause it not to bitch
about result models.
when a result has not been released yet it will be displayed.